feat: implement useDeleteRouteSchema hook and integrate into RemoveRoute component
