+++ packages/core/client/src/route-switch/antd/admin-layout/index.tsx
@@ -34,10 +34,10 @@ import {
   useMenuDragEnd,
   useParseURLAndParams,
   useRequest,
+  useRouterBasename,
   useSchemaInitializerRender,
   useSystemSettings,
   useToken,
-  useRouterBasename,
 } from '../../../';
 import {
   CurrentPageUidContext,
@@ -46,22 +46,22 @@ import {
   IsSubPageClosedByPageMenuProvider,
   useCurrentPageUid,
   useLocationNoUpdate,
+  useNavigateNoUpdate,
 } from '../../../application/CustomRouterContextProvider';
 import { Plugin } from '../../../application/Plugin';
+import { navigateWithinSelf } from '../../../block-provider/hooks';
 import { AppNotFound } from '../../../common/AppNotFound';
 import { withTooltipComponent } from '../../../hoc/withTooltipComponent';
+import { useEvaluatedExpression } from '../../../hooks/useParsedValue';
 import { menuItemInitializer } from '../../../modules/menu/menuItemInitializer';
 import { useMenuTranslation } from '../../../schema-component/antd/menu/locale';
+import { VariableScope } from '../../../variables/VariableScope';
 import { KeepAlive, useKeepAlive } from './KeepAlive';
 import { NocoBaseDesktopRoute, NocoBaseDesktopRouteType } from './convertRoutesToSchema';
 import { MenuSchemaToolbar, ResetThemeTokenAndKeepAlgorithm } from './menuItemSettings';
 import { userCenterSettings } from './userCenterSettings';
-import { navigateWithinSelf } from '../../../block-provider/hooks';
-import { useNavigateNoUpdate } from '../../../application/CustomRouterContextProvider';
-import { VariableScope } from '../../../variables/VariableScope';
-import _ from 'lodash';
-import { useEvaluatedExpression } from '../../../hooks/useParsedValue';
 
+export * from './useDeleteRouteSchema';
 export { KeepAlive, NocoBaseDesktopRouteType, useKeepAlive };
 
 export const NocoBaseRouteContext = createContext<NocoBaseDesktopRoute | null>(null);
