+++ packages/core/client/src/route-switch/antd/admin-layout/menuItemSettings.tsx
@@ -41,6 +41,7 @@ import {
   SchemaSettingsSwitchItem,
 } from '../../../schema-settings/SchemaSettings';
 import { NocoBaseDesktopRoute } from './convertRoutesToSchema';
+import { useDeleteRouteSchema } from './useDeleteRouteSchema';
 
 const components = { TreeSelect };
 
@@ -113,6 +114,7 @@ export const RemoveRoute: FC = () => {
   const { t } = useTranslation();
   const { modal } = App.useApp();
   const { deleteRoute } = useNocoBaseRoutes();
+  const { deleteRouteSchema } = useDeleteRouteSchema();
   const currentRoute = useCurrentRoute();
   const { allAccessRoutes } = useAllAccessDesktopRoutes();
   const navigate = useNavigateNoUpdate();
@@ -128,7 +130,9 @@ export const RemoveRoute: FC = () => {
           content: t('Are you sure you want to delete it?'),
           onOk: async () => {
             // 删除对应菜单的路由
-            currentRoute?.id != null && (await deleteRoute(currentRoute.id));
+            if (currentRoute?.id != null) {
+              await Promise.all([deleteRoute(currentRoute.id), deleteRouteSchema(currentRoute.schemaUid)]);
+            }
 
             if (
               currentPageUid !== currentRoute?.schemaUid &&
